refactor/#75 : User 데이터 클래스를 core 모듈로 이동
