+++ data/auth/src/main/java/com/boostcamp/mapisode/auth/UserMapper.kt
@@ -1,5 +1,6 @@
 package com.boostcamp.mapisode.auth
 
+import com.boostcamp.mapisode.model.User
 import com.google.firebase.auth.AuthResult
 import javax.inject.Inject
