+++ feature/login/build.gradle.kts
@@ -7,6 +7,7 @@ android {
 }
 
 dependencies {
+	implementation(projects.core.model)
 	implementation(projects.domain.auth)
 	implementation(platform(libs.firebase.bom))
 	implementation(libs.bundles.firebase)
